refactor: change split from main/read/write/manifest to main/image/manifest
