+++ backend/src/photosite_backend/tests/test_manifest.py
@@ -1,6 +1,7 @@
+import json
 from pathlib import Path
 
-from photosite_backend.manifest import generate_manifest_entry
+from photosite_backend.manifest import generate_manifest_entry, write_manifest
 from photosite_backend.tests import create_test_datafile
 
 
@@ -17,3 +18,15 @@ def test_generate_manifest_entry(tmp_path: Path):
             "DateTime": "2019:10:22 18:00:27",
         },
     }
+
+
+def test_write_manifest(tmpdir):
+    manifest = {
+        "version": 1,
+        "images": {"asdf": {"foo": "bar"}, "hjkl": {"baz": "qux"}},
+    }
+
+    manifest_path = write_manifest(tmpdir, manifest)
+
+    with manifest_path.open("r") as file:
+        assert json.load(file) == manifest
